feature: add social login with torus, add gooddollar wc button
