+++ package.json
@@ -54,6 +54,7 @@
         "@tailwindcss/line-clamp": "^0.2.0",
         "@tailwindcss/postcss7-compat": "^2.0.4",
         "@tailwindcss/typography": "^0.3.1",
+        "@toruslabs/torus-embed": "^1.38.5",
         "@types/bn.js": "^5.1.0",
         "@types/graphql": "^14.5.0",
         "@types/jest": "^25.2.1",
