+++ src/connectors/modules/CustomWc/helpers/icons/gooddollar/icon.js
@@ -1,0 +1,15 @@
+const GdIcon = `
+<svg width="40" height="40" viewBox="0 0 40 40" fill="none" xmlns="http://www.w3.org/2000/svg">
+<g clip-path="url(#clip0_1095_57377)">
+<circle cx="20" cy="20" r="20" fill="white"/>
+<path d="M40 20C40 8.95419 31.0458 0 19.9996 0C8.95512 0 0 8.95428 0 20.0001C0 31.046 8.95512 40 19.9996 40C31.0458 40 40 31.0457 40 20.0001V20ZM29.9793 10.5259L25.929 14.5764C25.929 14.5758 25.928 14.5758 25.928 14.5749L24.8746 15.6277C24.2685 15.0184 23.5476 14.5354 22.7536 14.2066C21.9596 13.8778 21.1083 13.7097 20.2489 13.7121C18.5147 13.714 16.8521 14.4038 15.6258 15.6301C14.3996 16.8564 13.7098 18.5191 13.708 20.2533C13.7097 21.9877 14.3993 23.6506 15.6256 24.8771C16.8519 26.1035 18.5146 26.7934 20.2489 26.7953C21.7294 26.7934 23.1655 26.2901 24.3234 25.3676C25.4813 24.4451 26.2927 23.1576 26.6254 21.715H18.7872L21.7095 18.7924H29.5988C29.6742 19.2758 29.7121 19.7642 29.7121 20.2533C29.7095 22.7624 28.7116 25.168 26.9375 26.9423C25.1634 28.7166 22.758 29.7147 20.2489 29.7176C19.0066 29.7184 17.7763 29.4737 16.6288 28.9977C15.4813 28.5217 14.4391 27.8236 13.5621 26.9437L12.1845 28.3212H8.04896L11.8193 24.5534C11.1381 23.2226 10.7837 21.7487 10.7854 20.2538C10.7882 17.7447 11.7862 15.3392 13.5603 13.5649C15.3345 11.7907 17.7399 10.7927 20.2489 10.7899C21.7448 10.7867 23.2199 11.14 24.5521 11.8203L25.8466 10.5259H29.9793Z" fill="#081C3E"/>
+</g>
+<defs>
+<clipPath id="clip0_1095_57377">
+<rect width="40" height="40" fill="white"/>
+</clipPath>
+</defs>
+</svg>
+`
+
+export default GdIcon
