+++ src/connectors/modules/CustomWc/helpers/icons/zengo/icon.js
