+++ src/connectors/modules/CustomWc/helpers/isMobile.ts
